separate pydantic models from cloud_server_model.py
