+++ .idea/misc.xml
@@ -3,5 +3,5 @@
   <component name="Black">
     <option name="sdkName" value="Poetry (timeweb-sdk) (2)" />
   </component>
-  <component name="ProjectRootManager" version="2" project-jdk-name="Poetry (timeweb-sdk)" project-jdk-type="Python SDK" />
+  <component name="ProjectRootManager" version="2" project-jdk-name="Poetry (timeweb-sdk) (2)" project-jdk-type="Python SDK" />
 </project>
