+++ .idea/timeweb-sdk.iml
@@ -6,7 +6,7 @@
       <sourceFolder url="file://$MODULE_DIR$/tests" isTestSource="true" />
       <excludeFolder url="file://$MODULE_DIR$/dist" />
     </content>
-    <orderEntry type="jdk" jdkName="Poetry (timeweb-sdk)" jdkType="Python SDK" />
+    <orderEntry type="jdk" jdkName="Poetry (timeweb-sdk) (2)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
   <component name="PyNamespacePackagesService">
