+++ pyproject.toml
@@ -10,7 +10,8 @@ readme = "README.md"
 requires-python = ">=3.13"
 dependencies = [
     "httpx (>=0.28.1,<0.29.0)",
-    "pydantic (>=2.11.1,<3.0.0)"
+    "pydantic (>=2.11.1,<3.0.0)",
+    "annotated-types (>=0.7.0,<0.8.0)"
 ]
 
 [project.urls]
@@ -23,6 +24,8 @@ requires-poetry = ">=2.0"
 
 [tool.poetry.group.dev.dependencies]
 ruff = "^0.9.10"
+pytest = "^8.3.5"
+respx = "^0.22.0"
 
 [tool.ruff]
 indent-width = 4
