+++ src/timeweb_sdk/models/__init__.py
@@ -1,3 +1,1 @@
-from .cloud_server_model import CloudServerModel
-
-__all__ = ["CloudServerModel"]
+from .cloud_server_models import *
