+++ src/timeweb_sdk/models/cloud_server_models/image_model.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel, Field
+from typing import List
+
+
+class ImageModel(BaseModel):
+    id: str
+    name: str
+    is_custom: bool
