+++ src/timeweb_sdk/models/cloud_server_models/os_model.py
@@ -1,0 +1,8 @@
+from pydantic import BaseModel, Field
+from typing import List
+
+
+class OSModel(BaseModel):
+    id: int
+    name: str
+    version: str
