+++ src/timeweb_sdk/models/cloud_server_models/software_model.py
@@ -1,0 +1,7 @@
+from pydantic import BaseModel, Field
+from typing import List
+
+
+class SoftwareModel(BaseModel):
+    id: int = Field(default=None)
+    name: str = Field(default=None)
